requirements.txt
biopython==1.80
click==8.1.3
ete3==3.1.2
matplotlib==3.7.0
numpy==1.24.1
wget==3.2
